arcana-cards: Tier-2 full-pool re-run, --force land (fidelity upgrade)
Second full-pool SingleEffectSpell production run with the Tier-2
prompt (B.1 rich target filters + Effect::Sacrifice; B.2 script::
all_players/opponents; B.3 Effect::DelayedAction). 27 batched
agent calls of 50-100 cards each generated the full 2,950-pool;
verify_dir --batch: 2,102 passed (71%) — L1 109, L2 6, L3-stub 733.
Pass-rate dropped from Tier-1 77% because the dynamic-literal lint
now correctly quarantines literal stand-ins agents previously shipped
as PARTIAL; the audit's CORRECT 30%→50% is the real quality gain.

Landed --force --apply: 2,098 written (overwriting existing where
the Tier-2 version passes; ~139 brand-new slugs); 4 SKIP from the C1
name-collision guard (seed-card slugs the sampler regenerates under
non-canonical sets — Counterspell, Lightning Bolt, Murder, Preordain —
correctly held back). register_all regenerated by the land hook (C1
wiring). Net: 2056 modified + 139 added = 2,195 catalog files
touched; +40,241 / −35,764 lines (Tier-2 resolvers are richer).

Validation: cargo check -p arcana-cards clean (warnings only);
arcana-cards 5/5 — incl. the C1 standing test (full ~3,200-card
register_all builds, no duplicate-name panic, registry intact);
seed_integration 108/108; verify_catalog still 1019/1019 clean.

Per the 120-card B re-pilot audit, the landed catalog's spell
fidelity should now be ~50% fully-correct (vs ~30% pre-Tier-2),
~46% partial-functional, ~4% wrong, 0% stub.

+++ arcana-cards/src/ddj/ghoul_s_feast.rs
@@ -1,16 +1,17 @@
-//! Ghoul's Feast — `{1}{B}` instant. "Target creature gets +X/+0 until end of
-//! turn, where X is the number of creature cards in your graveyard."
+//! Ghoul's Feast — `{1}{B}` instant. "Target creature gets +X/+0
+//! until end of turn, where X is the number of creature cards in your
+//! graveyard."
 
 use arcana_core::effects::Effect;
 use arcana_core::layers::Duration;
 use arcana_core::mana::ManaCost;
 use arcana_core::objects::Characteristics;
 use arcana_core::registry::{CardDefinition, CardRegistry, SpellAbilityDef};
+use arcana_core::script;
 use arcana_core::stack::StackEntry;
 use arcana_core::state::GameState;
 use arcana_core::targets::{ObjectFilter, TargetChoice, TargetRequirement};
 use arcana_core::types::{CardId, ColorSet, TypeLine};
-use arcana_core::script;
 
 pub fn register(reg: &mut CardRegistry) -> CardId {
     let name = reg.interner_mut().intern("Ghoul's Feast");
@@ -22,13 +23,12 @@ pub fn register(reg: &mut CardRegistry) -> CardId {
         ..Default::default()
     };
     reg.register(
-        CardDefinition::new(name, chars)
-            .with_spell_ability(SpellAbilityDef {
-                text: "Target creature gets +X/+0 until end of turn, where X is the number of creature cards in your graveyard.".into(),
-                target_requirements: vec![TargetRequirement::target_creature()],
-                modal: None,
-                effect: resolve,
-            }),
+        CardDefinition::new(name, chars).with_spell_ability(SpellAbilityDef {
+            text: "Target creature gets +X/+0 until end of turn, where X is the number of creature cards in your graveyard.".into(),
+            target_requirements: vec![TargetRequirement::target_creature()],
+            modal: None,
+            effect: resolve,
+        }),
     )
 }
 
@@ -39,10 +39,15 @@ fn resolve(
 ) -> Vec<Effect> {
     let Some(target) = entry.targets.targets.first() else { return Vec::new(); };
     let TargetChoice::Object(id) = target else { return Vec::new(); };
-    let x = script::graveyard_matching(state, &ObjectFilter::creature(), entry.controller, entry.controller) as i32;
+    let x = script::graveyard_matching(
+        state,
+        &ObjectFilter::creature(),
+        entry.controller,
+        entry.controller,
+    );
     vec![Effect::Pump {
         target: *id,
-        power: x,
+        power: x as i32,
         toughness: 0,
         duration: Duration::EndOfTurn,
         keywords: vec![],
